test(mocha): simplify few configurations (#6629)
* test(mocha): simplify few configurations

* build(ci): expand node.js version matrix
