+++ .github/workflows/ci_main.yml
@@ -10,7 +10,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        node: [ '12', '13', '14', '15' ]
+        node: [ '12', '13', '14', '15', '16' ]
 
     name: Node ${{ matrix.node }} build
