+++ spec/helpers/testScheduler-ui.ts
@@ -1,6 +1,5 @@
 import * as _ from 'lodash';
 import * as chai from 'chai';
-import * as sinonChai from 'sinon-chai';
 import * as marble from './marble-testing';
 import { TestScheduler } from 'rxjs/testing';
 
@@ -8,45 +7,6 @@ import { TestScheduler } from 'rxjs/testing';
 const commonInterface = require('mocha/lib/interfaces/common');
 const escapeRe = require('escape-string-regexp');
 //tslint:enable:no-var-requires no-require-imports
-
-/** Polyfill requestAnimationFrame for testing animationFrame scheduler in Node */
-// http://paulirish.com/2011/requestanimationframe-for-smart-animating/
-// http://my.opera.com/emoller/blog/2011/12/20/requestanimationframe-for-smart-er-animating
-
-// requestAnimationFrame polyfill by Erik Möller. fixes from Paul Irish and Tino Zijdel
-
-// MIT license
-
-(function(this: any, window: any) {
-  window = window || this;
-  let lastTime = 0;
-  const vendors = ['ms', 'moz', 'webkit', 'o'];
-  for (let x = 0; x < vendors.length && !window.requestAnimationFrame; ++x) {
-      window.requestAnimationFrame = window[vendors[x] + 'RequestAnimationFrame'];
-      window.cancelAnimationFrame = window[vendors[x] + 'CancelAnimationFrame']
-                                 || window[vendors[x] + 'CancelRequestAnimationFrame'];
-  }
-
-  if (!window.requestAnimationFrame) {
-      window.requestAnimationFrame = (callback: Function, element: any) => {
-          const currTime = new Date().getTime();
-          const timeToCall = Math.max(0, 16 - (currTime - lastTime));
-          const id = window.setTimeout(() => { callback(currTime + timeToCall); },
-            timeToCall);
-          lastTime = currTime + timeToCall;
-          return id;
-      };
-  }
-
-  if (!window.cancelAnimationFrame) {
-      window.cancelAnimationFrame = (id: number) => {
-          clearTimeout(id);
-      };
-  }
-}(global));
-
-//setup sinon-chai
-chai.use(sinonChai);
 
 declare const module: any, global: any, Suite: any, Test: any;
 
@@ -56,12 +16,16 @@ if (global && !(typeof window !== 'undefined')) {
   global.Test = global.mocha.Test;
 }
 
-//mocha creates own global context per each test suite, simple patching to global won't deliver its context into test cases.
-//this custom interface is just mimic of existing one amending test scheduler behavior previously test-helper does via global patching.
-module.exports = function(suite: any) {
+/**
+ * mocha creates own global context per each test suite, simple patching to global won't deliver its context into test cases.
+ * this custom interface is just mimic of existing one amending test scheduler behavior previously test-helper does via global patching.
+ * 
+ * @deprecated This ui is no longer actively used. Will be removed after migrating remaining tests uses this.
+ */
+module.exports = function (suite: any) {
   const suites = [suite];
 
-  suite.on('pre-require', function(context: any, file: any, mocha: any) {
+  suite.on('pre-require', function (context: any, file: any, mocha: any) {
     const common = (<any>commonInterface)(suites, context);
 
     context.before = common.before;
@@ -86,7 +50,7 @@ module.exports = function(suite: any) {
      * and/or tests.
      */
 
-    context.describe = context.context = function(title: any, fn: any) {
+    context.describe = context.context = function (title: any, fn: any) {
       const suite = (<any>Suite).create(suites[0], title);
       suite.file = file;
       suites.unshift(suite);
@@ -99,7 +63,7 @@ module.exports = function(suite: any) {
      * Pending describe.
      */
 
-    context.xdescribe = context.xcontext = context.describe.skip = function(title: any, fn: any) {
+    context.xdescribe = context.xcontext = context.describe.skip = function (title: any, fn: any) {
       const suite = (<any>Suite).create(suites[0], title);
       suite.pending = true;
       suites.unshift(suite);
@@ -111,7 +75,7 @@ module.exports = function(suite: any) {
      * Exclusive suite.
      */
 
-    context.describe.only = function(title: any, fn: any) {
+    context.describe.only = function (title: any, fn: any) {
       const suite = context.describe(title, fn);
       mocha.grep(suite.fullTitle());
       return suite;
@@ -127,9 +91,9 @@ module.exports = function(suite: any) {
         }
         return value;
       })
-      .replace(/\\"/g, '"')
-      .replace(/\\t/g, '\t')
-      .replace(/\\n/g, '\n');
+        .replace(/\\"/g, '"')
+        .replace(/\\t/g, '\t')
+        .replace(/\\n/g, '\n');
     }
 
     function deleteErrorNotificationStack(marble: any) {
@@ -174,7 +138,7 @@ module.exports = function(suite: any) {
      * acting as a thunk.
      */
 
-    const it = context.it = context.specify = function(title: any, fn: any) {
+    const it = context.it = context.specify = function (title: any, fn: any) {
       context.rxTestScheduler = null;
       let modified = fn;
 
@@ -205,7 +169,7 @@ module.exports = function(suite: any) {
      * Exclusive test-case.
      */
 
-    context.it.only = function(title: any, fn: any) {
+    context.it.only = function (title: any, fn: any) {
       const test = it(title, fn);
       const reString = '^' + (<any>escapeRe)(test.fullTitle()) + '$';
       mocha.grep(new RegExp(reString));
@@ -216,14 +180,14 @@ module.exports = function(suite: any) {
      * Pending test case.
      */
 
-    context.xit = context.xspecify = context.it.skip = function(title: string) {
+    context.xit = context.xspecify = context.it.skip = function (title: string) {
       context.it(title);
     };
 
     /**
      * Number of attempts to retry.
      */
-    context.it.retries = function(n: number) {
+    context.it.retries = function (n: number) {
       context.retries(n);
     };
   });
