+++ spec/support/.mocharc.js
@@ -1,5 +1,5 @@
 module.exports = {
-  require: ['spec/support/mocha-path-mappings.js', 'dist/spec/helpers/polyfills.js', 'dist/spec/helpers/testScheduler-ui.js'],
+  require: ['spec/support/mocha-path-mappings.js', 'dist/spec/helpers/setup.js'],
   ui: ['dist/spec/helpers/testScheduler-ui.js'],
   reporter: 'dot',
   timeout: 5000,
